WEB-1491: Move modal actions and fix tests
