+++ web/app/containers/app/actions.test.js
@@ -5,17 +5,21 @@
 /* eslint-env jest */
 
 import {fromJS} from 'immutable'
-import * as appActions from './actions'
 import {setPageFetchError, clearPageFetchError} from 'progressive-web-sdk/dist/store/offline/actions'
 import {OFFLINE_MODAL} from '../../modals/constants'
 import {OFFLINE_ASSET_URL} from './constants'
 import {UI_NAME} from 'progressive-web-sdk/dist/analytics/data-objects/'
+jest.mock('../../modals/actions')
+import {closeModal} from '../../modals/actions'
+import {checkIfOffline} from './actions'
+
 
 let realFetch
 beforeAll(() => {
     realFetch = global.fetch
     global.fetch = jest.fn()
     global.fetch.mockReturnValue(Promise.resolve())
+
 })
 
 afterAll(() => {
@@ -27,7 +31,7 @@ test('checkIfOffline dispatches setPageFetchError if network request fails', () 
     global.fetch.mockReturnValueOnce(Promise.reject(new TypeError('failed to fetch')))
 
     const fakeDispatch = jest.fn()
-    const thunk = appActions.checkIfOffline()
+    const thunk = checkIfOffline()
 
     return thunk(fakeDispatch)
         .then(() => {
@@ -56,7 +60,7 @@ test('checkIfOffline dispatches setPageFetchError if it receives modified JSON f
     const fakeGetState = () => fromJS({
         modals: undefined
     })
-    const thunk = appActions.checkIfOffline()
+    const thunk = checkIfOffline()
 
     return thunk(fakeDispatch, fakeGetState)
         .then(() => {
@@ -85,8 +89,8 @@ test('checkIfOffline clears offline modal and page fetch errors when it receives
             [OFFLINE_MODAL]: true
         })
     })
-    const thunk = appActions.checkIfOffline()
-    appActions.closeModal = jest.fn()  // eslint-disable-line
+    // closeModal = jest.fn()  // eslint-disable-line
+    const thunk = checkIfOffline()
 
     return thunk(fakeDispatch, fakeGetState)
         .then(() => {
@@ -95,7 +99,6 @@ test('checkIfOffline clears offline modal and page fetch errors when it receives
 
             expect(fakeDispatch).toHaveBeenCalledTimes(2)
             expect(fakeDispatch.mock.calls[0][0]).toEqual(clearPageFetchError())
-            expect(fakeDispatch).toBeCalled()
-            expect(appActions.closeModal).toHaveBeenCalledWith(OFFLINE_MODAL, UI_NAME.offline)
+            expect(closeModal).toHaveBeenCalledWith(OFFLINE_MODAL, UI_NAME.offline)
         })
 })
