+++ web/app/containers/cart/partials/cart-summary.jsx
@@ -7,7 +7,7 @@ import {connect} from 'react-redux'
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import * as cartSelectors from 'progressive-web-sdk/dist/store/cart/selectors'
 import {CART_ESTIMATE_SHIPPING_MODAL} from '../../../modals/constants'
-import {openModal} from '../../app/actions'
+import {openModal} from '../../../modals/actions'
 import {getSelectedShippingLabel, getPostcode} from '../../../store/checkout/shipping/selectors'
 import {getCheckoutShippingURL} from '../../app/selectors'
 import {removePromoCode} from '../actions'
