+++ web/app/containers/checkout-confirmation/actions.js
@@ -8,7 +8,7 @@ import {createAction} from 'progressive-web-sdk/dist/utils/action-creation'
 import {createSelector} from 'reselect'
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import {addNotification, removeAllNotifications} from 'progressive-web-sdk/dist/store/notifications/actions'
-import {openModal} from '../app/actions'
+import {openModal} from '../../modals/actions'
 import * as shippingSelectors from '../../store/checkout/shipping/selectors'
 import * as formSelectors from '../../store/form/selectors'
 import {getBillingAddress} from '../../store/checkout/billing/selectors'
