+++ web/app/containers/header/container.jsx
@@ -10,7 +10,7 @@ import classnames from 'classnames'
 
 import * as headerActions from './actions'
 import * as miniCartActions from '../../modals/mini-cart/actions'
-import {openModal} from '../app/actions'
+import {openModal} from '../../modals/actions'
 import {NAVIGATION_MODAL, MORE_MENU} from '../../modals/constants'
 import * as selectors from './selectors'
 import {getCartSummaryCount} from 'progressive-web-sdk/dist/store/cart/selectors'
