+++ web/app/containers/product-details/actions.js
@@ -22,6 +22,7 @@ import {updateWishlistItem} from 'progressive-web-sdk/dist/integration-manager/a
 import {addNotification} from 'progressive-web-sdk/dist/store/notifications/actions'
 import {UI_NAME} from 'progressive-web-sdk/dist/analytics/data-objects/'
 import {PRODUCT_DETAILS_ITEM_ADDED_MODAL} from '../../modals/constants'
+import {closeModal, openModal} from '../../modals/actions'
 import * as appActions from '../app/actions'
 
 export const setIsWishlistAdded = createAction('Set is wishlist added', ['isWishlistAdded'])
@@ -38,13 +39,13 @@ export const addToCartStarted = createAction('Add to cart started')
 export const addToCartComplete = createAction('Add to cart complete')
 
 export const goToCheckout = () => (dispatch) => {
-    dispatch(appActions.closeModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.addToCart))
+    dispatch(closeModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.addToCart))
     dispatch(appActions.goToCheckout())
 }
 
 export const goToWishlist = () => (dispatch, getState) => {
     dispatch(setIsWishlistAdded(false))
-    dispatch(appActions.closeModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.wishlist))
+    dispatch(closeModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.wishlist))
 
     browserHistory.push({
         pathname: getWishlistURL(getState())
@@ -100,7 +101,7 @@ export const submitCartForm = (formValues) => (dispatch, getStore) => {
     return dispatch(itemIdMatch ? updateCartItem(itemIdMatch[1], qty, productId, variant) : addToCart(productId, qty, variant))
         .then(() => {
             dispatch(setIsWishlistAdded(false))
-            return dispatch(appActions.openModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.addToCart))
+            return dispatch(openModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.addToCart))
         })
         .catch((error) => {
             console.error('Error adding to cart', error)
@@ -150,7 +151,7 @@ export const addToWishlist = (quantity) => (dispatch, getState) => {
     return dispatch(addItemToWishlist(productID, window.location.href, quantity))
         .then(() => {
             dispatch(setIsWishlistAdded(true))
-            return dispatch(appActions.openModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.wishlist))
+            return dispatch(openModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.wishlist))
         })
         .catch((error) => {
             if (/Failed to fetch|Add Request Failed|Unable to add item/i.test(error.message)) {
@@ -172,6 +173,6 @@ export const updateItemInWishlist = (quantity) => (dispatch, getState) => {
     return dispatch(updateWishlistItem(itemId, wishlistId, quantity))
         .then(() => {
             dispatch(setIsWishlistAdded(true))
-            return dispatch(appActions.openModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.wishlist))
+            return dispatch(openModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.wishlist))
         })
 }
