+++ web/app/modals/account-add-address/container.jsx
@@ -6,7 +6,7 @@ import React from 'react'
 import {connect} from 'react-redux'
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import {ACCOUNT_ADDRESS_MODAL} from '../constants'
-import {closeModal} from '../../containers/app/actions'
+import {closeModal} from './actions'
 import {isModalOpen} from 'progressive-web-sdk/dist/store/modals/selectors'
 import AccountAddressReduxForm from '../../containers/account-address/partials/address-form'
 import AddAddressHeader from './partials/add-address-header'
