+++ web/app/modals/cart-wishlist/container.jsx
@@ -7,7 +7,7 @@ import {connect} from 'react-redux'
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import {getAssetUrl} from 'progressive-web-sdk/dist/asset-utils'
 import {CART_WISHLIST_MODAL} from '../constants'
-import {closeModal} from '../../containers/app/actions'
+import {closeModal} from '../actions'
 import {setIsWishlistComplete} from '../../containers/cart/actions'
 import {isModalOpen} from 'progressive-web-sdk/dist/store/modals/selectors'
 import {getIsLoggedIn} from 'progressive-web-sdk/dist/store/user/selectors'
