+++ web/app/modals/checkout-confirmation/container.jsx
@@ -4,7 +4,7 @@
 
 import React from 'react'
 import {CHECKOUT_CONFIRMATION_MODAL} from '../constants'
-import {closeModal} from '../../containers/app/actions'
+import {closeModal} from './actions'
 import {connect} from 'react-redux'
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import {getAssetUrl} from 'progressive-web-sdk/dist/asset-utils'
